Allow users to pass the encoding tokenizer
